Fixed TLSCACert typo and other misc. issues
Fixed the config parameter used for TLSCACertFile in server.go so that
it matches the documentation (and makes more logical sense) as well as
corrected the associated error message so that it was not a duplicate of
the one for TLSKeyFile.

Removed some trailing whitespace characters identified in VIM which
were present in some files.

Removed out-of-date precompiled binary files from the source tree.
Recompiled the .tar.gz file of the source located in the build directory
with the changes made in this commit.